Move ReuseInfo to copyright
ReuseInfo will depend on CopyrightNotice, so it makes sense to move it.
