+++ src/reuse/_annotate.py
@@ -22,7 +22,6 @@ from typing import IO, Optional, Type, cast
 from jinja2 import Environment, FileSystemLoader, Template
 from jinja2.exceptions import TemplateNotFound
 
-from . import ReuseInfo
 from ._util import _determine_license_suffix_path
 from .comment import (
     NAME_STYLE_MAP,
@@ -30,6 +29,7 @@ from .comment import (
     EmptyCommentStyle,
     get_comment_style,
 )
+from .copyright import ReuseInfo
 from .exceptions import CommentCreateError, MissingReuseInfoError
 from .extract import contains_reuse_info, detect_line_endings
 from .header import add_new_header, find_and_replace_header
