+++ src/reuse/copyright.py
@@ -13,8 +13,9 @@ from enum import Enum, unique
 from io import StringIO
 from typing import Any, Iterable, Literal, NewType, Optional, Union, cast
 
+from boolean.boolean import Expression
+
 from .exceptions import CopyrightNoticeParseError, YearRangeParseError
-from .extract import _COPYRIGHT_PATTERNS  # TODO: Get rid of this import.
 from .i18n import _
 
 _LOGGER = logging.getLogger(__name__)
@@ -488,6 +489,99 @@ class CopyrightNotice:
         return str(self)
 
 
+class SourceType(Enum):
+    """
+    An enumeration representing the types of sources for license information.
+    """
+
+    #: A .license file containing license information.
+    DOT_LICENSE = "dot-license"
+    #: A file header containing license information.
+    FILE_HEADER = "file-header"
+    #: A .reuse/dep5 file containing license information.
+    DEP5 = "dep5"
+    #: A REUSE.toml file containing license information.
+    REUSE_TOML = "reuse-toml"
+
+
+# TODO: In Python 3.10+, add kw_only=True
+@dataclass(frozen=True)
+class ReuseInfo:
+    """Simple dataclass holding licensing and copyright information"""
+
+    spdx_expressions: set[Expression] = field(default_factory=set)
+    copyright_lines: set[str] = field(default_factory=set)
+    contributor_lines: set[str] = field(default_factory=set)
+    path: Optional[str] = None
+    source_path: Optional[str] = None
+    source_type: Optional[SourceType] = None
+
+    def _check_nonexistent(self, **kwargs: Any) -> None:
+        nonexistent_attributes = set(kwargs) - set(self.__dict__)
+        if nonexistent_attributes:
+            raise KeyError(
+                f"The following attributes do not exist in"
+                f" {self.__class__}: {', '.join(nonexistent_attributes)}"
+            )
+
+    def copy(self, **kwargs: Any) -> "ReuseInfo":
+        """Return a copy of ReuseInfo, replacing the values of attributes with
+        the values from *kwargs*.
+        """
+        self._check_nonexistent(**kwargs)
+        new_kwargs = {}
+        for key, value in self.__dict__.items():
+            new_kwargs[key] = kwargs.get(key, value)
+        return self.__class__(**new_kwargs)  # type: ignore
+
+    def union(self, value: "ReuseInfo") -> "ReuseInfo":
+        """Return a new instance of ReuseInfo where all set attributes are equal
+        to the union of the set in *self* and the set in *value*.
+
+        All non-set attributes are set to their values in *self*.
+
+        >>> one = ReuseInfo(copyright_lines={"Jane Doe"}, source_path="foo.py")
+        >>> two = ReuseInfo(copyright_lines={"John Doe"}, source_path="bar.py")
+        >>> result = one.union(two)
+        >>> print(sorted(result.copyright_lines))
+        ['Jane Doe', 'John Doe']
+        >>> print(result.source_path)
+        foo.py
+        """
+        new_kwargs = {}
+        for key, attr_val in self.__dict__.items():
+            if isinstance(attr_val, set) and (other_val := getattr(value, key)):
+                new_kwargs[key] = attr_val.union(other_val)
+            else:
+                new_kwargs[key] = attr_val
+        return self.__class__(**new_kwargs)  # type: ignore
+
+    def contains_copyright_or_licensing(self) -> bool:
+        """Either *spdx_expressions* or *copyright_lines* is non-empty."""
+        return bool(self.spdx_expressions or self.copyright_lines)
+
+    def contains_copyright_xor_licensing(self) -> bool:
+        """One of *spdx_expressions* or *copyright_lines* is non-empty."""
+        return bool(self.spdx_expressions) ^ bool(self.copyright_lines)
+
+    def contains_info(self) -> bool:
+        """Any field except *path*, *source_path* and *source_type* is
+        non-empty.
+        """
+        keys = {
+            key
+            for key in self.__dict__
+            if key not in ("path", "source_path", "source_type")
+        }
+        return any(self.__dict__[key] for key in keys)
+
+    def __bool__(self) -> bool:
+        return any(self.__dict__.values())
+
+    def __or__(self, value: "ReuseInfo") -> "ReuseInfo":
+        return self.union(value)
+
+
 # TODO: Everything below this comment will be removed in a future refactoring.
 
 
@@ -497,6 +591,10 @@ def merge_copyright_lines(copyright_lines: set[str]) -> set[str]:
 
     If a same statement uses multiple prefixes, use only the most frequent one.
     """
+    # pylint: disable=too-many-locals,import-outside-toplevel
+    # TODO: Remove this.
+    from .extract import _COPYRIGHT_PATTERNS  # TODO: Get rid of this import.
+
     # TODO: Rewrite this function. It's a bit of a mess.
     copyright_in = []
     for line in copyright_lines:
@@ -557,8 +655,12 @@ def make_copyright_line(
     """Given a statement, prefix it with ``SPDX-FileCopyrightText:`` if it is
     not already prefixed with some manner of copyright tag.
     """
+    # pylint: disable=import-outside-toplevel
     if "\n" in statement:
         raise RuntimeError(f"Unexpected newline in '{statement}'")
+
+    # TODO: Remove this.
+    from .extract import _COPYRIGHT_PATTERNS  # TODO: Get rid of this import.
 
     for pattern in _COPYRIGHT_PATTERNS:
         match = pattern.search(statement)
