+++ src/reuse/project.py
@@ -22,9 +22,9 @@ from typing import Collection, Iterator, NamedTuple, Optional, Type
 import attrs
 from binaryornot.check import is_binary
 
-from . import ReuseInfo
 from ._licenses import EXCEPTION_MAP, LICENSE_MAP
 from ._util import _determine_license_path, relative_from_root
+from .copyright import ReuseInfo
 from .covered_files import iter_files
 from .exceptions import (
     GlobalLicensingConflictError,
