+++ tests/test_global_licensing.py
@@ -13,7 +13,8 @@ from conftest import RESOURCES_DIRECTORY, posix
 from debian.copyright import Copyright
 from license_expression import LicenseSymbol
 
-from reuse import _LICENSING, ReuseInfo, SourceType
+from reuse import _LICENSING
+from reuse.copyright import ReuseInfo, SourceType
 from reuse.exceptions import (
     GlobalLicensingParseError,
     GlobalLicensingParseTypeError,
