+++ tests/test_project.py
@@ -20,7 +20,8 @@ import pytest
 from conftest import RESOURCES_DIRECTORY
 from license_expression import LicenseSymbol
 
-from reuse import _LICENSING, ReuseInfo, SourceType
+from reuse import _LICENSING
+from reuse.copyright import ReuseInfo, SourceType
 from reuse.covered_files import iter_files
 from reuse.exceptions import (
     GlobalLicensingConflictError,
